backend/pyproject.toml
[project]
name = "seaderr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "click>=8.2.1",
    "python-socketio>=5.13.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
build = ["hatch>=1.14.1"]


[tool.setuptools.packages.find]
where = ["src"]


[project.scripts]
seaderr = "seaderr.__main__:main"

[tool.uv]
package = true
